+++ extensions/default/src/utils/promptSaveReport.tsx
@@ -19,6 +19,7 @@ async function promptSaveReport({ servicesManager, commandsManager, extensionMan
   const isBackupSave = evt.isBackupSave === undefined ? evt.data.isBackupSave : evt.isBackupSave;
   const StudyInstanceUID = evt?.data?.StudyInstanceUID || ctx.trackedStudy;
   const SeriesInstanceUID = evt?.data?.SeriesInstanceUID;
+  const { displaySetInstanceUID } = evt.data ?? evt;
 
   const {
     trackedSeries,
@@ -86,6 +87,7 @@ async function promptSaveReport({ servicesManager, commandsManager, extensionMan
       SeriesInstanceUID,
       viewportId,
       isBackupSave,
+      displaySetInstanceUID,
     };
   } catch (error) {
     console.warn('Unable to save report', error);